Initial tests for v2_meta
